+++ src/gatsby-node.js
@@ -1,65 +1,3 @@
-import { gql, request } from "graphql-request";
-import readingTime from "reading-time";
+import { sourceNodes } from "./utils/source-nodes";
 
-const NODE_TYPE = "HashNodePost";
-
-exports.sourceNodes = async (
-  { actions, createNodeId, createContentDigest },
-  { username }
-) => {
-  // check username exists
-  if (!username) {
-    throw new Error("no username supplied");
-  }
-
-  const variables = {
-    username,
-  };
-
-  // query information
-  const query = gql`
-    query getPosts($username: String!) {
-      user(username: $username) {
-        publicationDomain
-        publication {
-          posts {
-            cuid
-            slug
-            title
-            brief
-            coverImage
-            dateAdded
-            _id
-            contentMarkdown
-          }
-        }
-      }
-    }
-  `;
-
-  const {
-    user: { publicationDomain, publication: { posts = [] } = {} },
-  } = await request("https://api.hashnode.com/", query, variables);
-
-  const { createNode } = actions;
-
-  posts.map((post) => {
-    const { _id, contentMarkdown = "" } = post;
-
-    const node = {
-      ...post,
-      readingTime: readingTime(contentMarkdown),
-      id: createNodeId(_id),
-      parent: null,
-      children: [],
-      internal: {
-        type: NODE_TYPE,
-        mediaType: `text/markdown`,
-        content: contentMarkdown,
-        contentDigest: createContentDigest(post),
-      },
-    };
-
-    createNode(node);
-  });
-};
+exports.sourceNodes = sourceNodes;
